accidentally overwriting 112M zip w/ all other produced zips (resulting in 40M xulrunner zip)
git-svn-id: http://svn.jboss.org/repos/jbosstools/trunk@23252 a97e2381-89e5-4abb-bab3-167db6db766c

+++ build/publish.sh
@@ -11,21 +11,25 @@ if [[ $DESTINATION == "" ]]; then DESTINATION="tools@filemgmt.jboss.org:/downloa
 # cleanup from last time
 rm -fr ${WORKSPACE}/site; mkdir -p ${WORKSPACE}/site/${JOB_NAME}
 
-for z in ${WORKSPACE}/sources/site/target/site.zip ${WORKSPACE}/sources/site/target/site_assembly.zip; do
-	if [[ -f $z ]]; then
-		#echo "$z ..."
-		# note the job name, build number, and build ID of the latest snapshot zip
-		echo "JOB_NAME = ${JOB_NAME}" > ${WORKSPACE}/site/${JOB_NAME}/JOB_NAME.txt
-		echo "BUILD_NUMBER = ${BUILD_NUMBER}" > ${WORKSPACE}/site/${JOB_NAME}/BUILD_NUMBER.txt
-		echo "BUILD_ID = ${BUILD_ID}" > ${WORKSPACE}/site/${JOB_NAME}/BUILD_ID.txt
-	
-		# unzip into workspace for publishing as unpacked site
-		unzip -u -o -q -d ${WORKSPACE}/site/${JOB_NAME}/ $z
-	
-		# copy into workspace for access by bucky aggregator (same name every time)
-		rsync -aq $z ${WORKSPACE}/site/${SNAPNAME}
-	fi
-done
+siteZip=${WORKSPACE}/sources/site/target/site_assembly.zip
+if [[ ! -f ${WORKSPACE}/sources/site/target/site_assembly.zip ]]; then
+	siteZip=${WORKSPACE}/sources/site/target/site.zip
+fi
+
+z=$siteZip
+if [[ -f $z ]]; then
+	#echo "$z ..."
+	# note the job name, build number, and build ID of the latest snapshot zip
+	echo "JOB_NAME = ${JOB_NAME}" > ${WORKSPACE}/site/${JOB_NAME}/JOB_NAME.txt
+	echo "BUILD_NUMBER = ${BUILD_NUMBER}" > ${WORKSPACE}/site/${JOB_NAME}/BUILD_NUMBER.txt
+	echo "BUILD_ID = ${BUILD_ID}" > ${WORKSPACE}/site/${JOB_NAME}/BUILD_ID.txt
+
+	# unzip into workspace for publishing as unpacked site
+	unzip -u -o -q -d ${WORKSPACE}/site/${JOB_NAME}/ $z
+
+	# copy into workspace for access by bucky aggregator (same name every time)
+	rsync -aq $z ${WORKSPACE}/site/${SNAPNAME}
+fi
 
 # if component zips exist, copy them too; first site.zip, then site_assembly.zip
 for z in $(find ${WORKSPACE}/sources/*/site/target -type f -name "site*.zip" | sort -r); do 
@@ -41,11 +45,13 @@ for z in $(find ${WORKSPACE}/sources/*/site/target -type f -name "site*.zip" | s
 done
 
 # if zips exist produced & renamed by ant script, copy them too
-for z in $(find ${WORKSPACE} -maxdepth 5 -mindepth 3 -name "*Update*.zip"); do 
-	#echo "$z ..."
-	unzip -u -o -q -d ${WORKSPACE}/site/${JOB_NAME}/ $z
-	rsync -aq $z ${WORKSPACE}/site/${SNAPNAME}
-done
+if [[ ! -f ${WORKSPACE}/site/${SNAPNAME} ]]; then
+	for z in $(find ${WORKSPACE} -maxdepth 5 -mindepth 3 -name "*Update*.zip"); do 
+		#echo "$z ..."
+		unzip -u -o -q -d ${WORKSPACE}/site/${JOB_NAME}/ $z
+		rsync -aq $z ${WORKSPACE}/site/${SNAPNAME}
+	done
+fi
 
 # get full build log and filter out Maven test failures
 bl=${WORKSPACE}/site/${JOB_NAME}/BUILDLOG.txt
